Estimation

State space model hyper parameters optimization routines.

Maximum Likelihood

StateSpace.maximum_likelihood! — Function

maximum_likelihood!(model; init=NamedTuple(), method=:collapsed, pen=NoPen(), ϵ_abs=1e-7, ϵ_rel=1e-4, max_iter=1000)

Maximum Likelihood (ML) algorithm to estimate the hyper parameters of a linear Gaussian State Space model as defined by model, results are stored in model. Penalized ML estimation is allowed through pen. If pen` = NoPen()`` the optimization routine is L-BFGS, whenpen≂̸ NoPen()` optimization is doen through ADMM.

Arguments

model::StateSpaceModel: state space model

init::NamedTuple : initial model parameters

method::Symbol : filtering method

pen::Penalization : penalization type

ϵ_abs::Real : absolute tolerance

ϵ_rel::Real : relative tolerance

max_iter::Integer : max number of iterations

Returns

ll::Real : log-likelihood value

Expectation-Maximization

StateSpace.em! — Function

em!(model, pen; init=NamedTuple(), ϵ=1e-4, max_iter=1000)

Expectation-Maximization (EM) algorithm to estimate the hyper parameters of a linear Gaussian State Space model as defined by model, storing the results in model.

Arguments

model::StateSpaceModel: state space model

pen::NamedTuple : penalization parameters

method::Symbol : filtering method

init::NamedTuple : initial model parameters

ϵ::Real : tolerance

max_iter::Integer : max number of iterations

Returns

ll::Real : log-likelihood value

StateSpace.ecm! — Function

ecm!(model, pen; init=NamedTuple(), ϵ=1e-4, max_iter=1000)

Expectation-Conditional Maximization (ECM) algorithm to estimate the hyper parameters of a linear Gaussian State Space model as defined by model, results are stored in model.

Arguments

model::StateSpaceModel: state space model

pen::NamedTuple : penalization parameters

method::Symbol : filtering method

init::NamedTuple : initial model parameters

ϵ::Real : tolerance

max_iter::Integer : max number of iterations

Returns

ll::Real : log-likelihood value