Models

A variety of model specifications in state space formulation.

Dynamic Factor Models

StateSpace.DynamicFactorModel — Type

DynamicFactorModel <: StateSpaceModel

Constructor for a dynamic factor model instance of the state space model type with hyperparameters Λ and ϕ, error specification, and mean specification.

StateSpace.DynamicNelsonSiegelModel — Type

DynamicNelsonSiegelModel <: StateSpaceModel

Constructor for a dynamic Nelson Siegel model instance of the state space model type with hyperparameters λ, ϕ, and error specifications

Mean Specifications

StateSpace.NoConstant — Type

NoConstant <: AbstractMeanModel

Constructor for a no constant mean specification instance of the mean model type.

StateSpace.Constant — Type

Constant <: AbstractMeanModel

Constructor for a constant mean specification instance of the mean model type, with hyperparameter μ.

StateSpace.Exogeneous — Type

Exogeneous <: AbstractMeanModel

Constructor for a exogeneous regressors mean model specification instance of the mean model type, with hyperparameter β and exogenous regressors X.

Error Specifications

StateSpace.Independent — Type

Independent <: AbstractErrorModel

Constructor for a independent error specification instance of the error model type, with hyperparameters σ and ω.

StateSpace.Idiosyncratic — Type

Idiosyncratic <: AbstractErrorModel

Constructor for a idiosyncratic error specification instance of the error model type, with hyperparameters Σ and Ω.

StateSpace.SpatialErrorModel — Type

SpatialErrorModel <: AbstractErrorModel

Constructor for a spatial error model specification instance of the error model type, with hyperparameter ρ, idiosyncratic error specification error, spatial weight matrix W, and group structure groups.

StateSpace.SpatialMovingAverageModel — Type

SpatialMovingAverageModel <: AbstractErrorModel

Constructor for a spatial moving average model specification instance of the error model type, with hyperparameter ρ, idiosyncratic error specification error, spatial weight matrix W, and group structure groups.